client/requirements.txt
flask-socketio
flask
grpcio
grpcio-tools
